Backports of r333636, r333637, r333640
r333636:
Fix for compile error in Unity 2018.3 with using TerrainData.SetTreeInstances
which isn't available until Unity. 2019.1+

r333637:
Fix for Rebuild using async version of the Recook after applying parm presets. It should
use the blocking Recook since we want to guarantee the Rebuild operation completes its full
cook.

r333640:
Fix for unity_script attribute specifying script to attach to generated terrain
not working for heightfield nodes. Reason was that the attribute class was required to be
Detail but for heightfields, the Detail attributes are automatically converted to Primitive class.
The fix supports unity_script regardless of attribute class.