Backports of r337473, r337474, r337476
r337473
Added PDG cook event message string handle (msgSH) to HAPI_PDG_EventInfo. This is set when
a cook event has a message (most times an error), which is then stored in the HAPI_String
cache which the user can look up later. The message string handle is valid until next PDG cook.
Bumped the HAPI version.

337474, 337476
PDG cook event messages and errors are now displayed on the PDGAssetLink UI.
These are the same errors that is shown on the node when PDG graph is cooked in Houdini.