Backports of 349583, 349593, 349599, 349812, 349821, 349953, 349966
r349583, r349593, r349599, r349812:
Fixes for validation:
-fixed slider UI when value is higher than max
-fixed horizontal line for UI separator
-can no longer create empty named shelf in Houdini Engine Tools Shelf
-handle Reset on HEU_HoudiniAssetRoot properly by re-assigning HEU_HoudiniAsset reference,
and reset params & rebuild HDA
-fixed Save HIP File cancel not displaying empty dialog

r349821:
Changed how the Houdini paths are written out when
HEU_HoudiniVersion.cs is generated.

r349953:
When an hda is converted to a prefab, a dialog message is displayed
to the user warning that it is not recommended, and a button is provided to
revert the prefab instance back to regular gameobject, and the new prefab
is removed. Also disables the hda prefab from being interacted with.
This is because the plugin does not support hda prefabs.

r349966:
Fixed up Houdini installation path lookup in registry for
Houdini Indie Steam on Windows.