Fixed a bug where recooking an asset may result in frozen locked nodes.
Seelan added the logic to push local attributes to Houdini in Revision 355479. This is for attribute paint and attribute edit functionality in Unity. However, we directly call SyncDirtyAttributesToHoudini(), which doesn't check if the attribute is actually dirty first. We only ever set an attributeStore to be dirty when doing a Unity-side Paint/Edit operation, so this is a risk free minimal change.