Fixed an issue where terrain layers were not being applied in PDG
See r373604:

"Unity plugin: fixed issue with the way that the plugin was retrieving the unity_hf_terrainlayer_file
attribute. The GetAttribute method will try every attribute owner, but since this volume prim attribute could be
transferred to the volume's vertices, it was causing the plugin to think something was wrong. This particular
code path will now call GetAttributeStrict, which will only try to retrieve the attribute for the requested
attribute owner, and will fail if it does not exist."